// web/src/shared/api/generated/model/organizationRole.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Cadence API
 * Project and work management for teams.
 * OpenAPI spec version: 1
 */

export type OrganizationRole = (typeof OrganizationRole)[keyof typeof OrganizationRole];

export const OrganizationRole = {
  Guest: 'Guest',
  Member: 'Member',
  Admin: 'Admin',
  Owner: 'Owner',
} as const;